Factor out BoxHandler.  Update HomePageTemplate.
